Initial the directories to be the base of module development
To develop a module in Magento 2, it requires many directories that need
to be correspond with the Magento 2 system.
    
These folders will be the basis of module development that can be
distributed to other developers to work together in parallel.
    
Note:
- Due to Git is not allow to commit the directory that has no any file,
file, tmp.txt, was used to be the dummy temporary file.

- Delete the file, tmp.txt, when develop the function or feature.