Change type of controls at setting page, from checkbox to select
Due to the automatic mechanism of the Magento 2, it will automatically
manages the control that the type is select.

Although the type of the control should be checkbox to deliver the
better user interface and experience and to remain the consistency with
other plugins but the current situation that the time to develop is
short and to implement the checkbox it requires more time to explore and
gain the knowledge.

Therefore, using the automatic mechanism of the Magento 2 to let it
manages the controls, it  required lower time to develop.