Create the charge with authorize and capture by using the key from DB
Create the charge with authorize and capture. The public key and secret
key will be used from the initialize of parent class.

Add a parameter, capture, and set the value to be true.

Although, without this parameter, capture, the charge will be captured
automatically by default but this adding parameter will be add more
intention to this charge creation. Others developer do not need to set
the assumption.