Implement new REST API to retrieve authorize_uri from an order
Because Magento use AJAX to communicate between checkout form (frontend) & placeOrder function (backend) by POST request to an payment-information api (and this API return an order id (int)).
So, to retrieve authorize_uri and redirect buyer to 3-D Secure page, we have to wait until the placeOrder process finished, which is we have to re-call to another API endpoint to retrieve OmiseCharge object again and do redirect at the frontend side.