Won't restore a quote if callback validation is failed
Magento has an issue that it cannot perform 'Order::registerCancellation()' and 'Session::restoreQuote()' to gether.
It will cause a product catalog quantity issue, both 'cancellation' and 'restoreQuote' will revert a quantity amount back to a product in the catalog, so, if we execute those functions together, we will get a wrong amount of quantity in a product catalog.