CreditCard Method: perform 'initialize' action if 3-D Secure payment is enabled.
Previously, order status flow when using 3-D Secure payment was, 'payment-review' > 'accept-payment' > 'processing' which is,
in fact, we should reserve that 'payment-review' for other situation and use 'pending-payment' instead due to that
there is no transaction at the time user is redirected to 3-D Secure page until user validate with OTP.

To correct the 3-D Secure payment behaviour, we need to rely on the payment 'initialize' method to set an order's status to be 'pending-payment'.
Note that there is no other way to set an order status to 'pending-payment' without using payment 'initialize' method.