Rename 'OmiseDataAssignObserver' to 'CreditCardDataObserver', make its name and property more relevant to the context.
Originally we named it as OmiseDataAssignObserver because we didn't prepare for the coming payment methods (Internet Banking, Alipay, Bill Payment, and so on) led to the irrelevant of class's name.
Since this class is taking care of only the 'credit card payment' data such as 'token' and 'card' (in the future once we finished 'save-card' feature).
So we should make its name more precise with its function.