Rename class 'PaymentCcTokenBuilder' to 'CreditCardBuilder'.
Regarding to the coming 'save-card' feature.
There will be a condition where the card's value will be assigned either as a 'token' or as a 'card_id' depending on what buyer choose from the front-end.
Meaning that, 'PaymentCcTokenBuilder' name will not reflect with the above condition

So here, I think to rename this class to make it conform with its functionality is important one.

Note. Originally, this class has been designed to only serve for 'card = token_id' purpose (we didn't design this class to support save card feature) therefore we need to update it.